memento/pyproject.toml
[project]
name = "memento"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud>=0.34.0",
    "google-cloud-texttospeech>=2.34.0",
    "langchain>=1.2.10",
    "langgraph>=1.0.10",
]
